--- a/src/contracts/SmartContract.sol
+++ b/src/contracts/SmartContract.sol
@@ -32,7 +32,7 @@ contract SmartContract {
     uint public nextUnsolvedPuzzlePart = 0;
     address public owner;
 
-    constructor() public payable {
+    constructor() public {
         owner = msg.sender;
     }