--- a/src/contracts/SmartContract.sol
+++ b/src/contracts/SmartContract.sol
@@ -23,16 +23,22 @@ contract SmartContract {
         _;
     }
 
+    // TODO - consider batching the events (in groups to reduce the fixed costs)
+    event Initialized( uint256 index, uint256 coins, uint256 upperBound, bytes32 prevPuzzleDetailsStorageHash);
+    event CommitmentSet(uint256 index, bytes32 commitment, bytes32 prevPuzzleCommitmentStorageHash);
+    event SolutionReceived(uint256 revIndex, bytes solution, bytes witness);
+
     address public helperID;
-    uint public startTime;
+    uint256 public startTime;
     Status public contractStatus = Status.Setup;
 
-    mapping(uint => PuzzlePart) public puzzleParts;
-    uint public amountOfPuzzleParts;
-    uint public nextUnsolvedPuzzlePart = 0;
+    bytes32 public puzzleDetailsStorageHash = bytes32(0);
+    bytes32 public puzzleCommitmentStorageHash = bytes32(0);
+
+    uint256 public amountOfPuzzleParts = 0;
     address public owner;
 
-    constructor() public {
+    constructor() payable public {
         owner = msg.sender;
     }
 
@@ -46,149 +52,107 @@ contract SmartContract {
         require(_coins.length > 0, "The length of the coins array should be greater than 0.");
         require(_helperID != address(0), "The helper ID should not be the zero address.");
         require(contractStatus == Status.Setup, "Contract setup is already completed.");
+        require(amountOfPuzzleParts == 0, "The amount of puzzle parts should be 0.");
 
-        if (amountOfPuzzleParts == 0) {
-            helperID = _helperID;
-            startTime = block.timestamp;
-            owner = msg.sender;
-        }
+        helperID = _helperID;
+        startTime = block.timestamp;
+        owner = msg.sender;
 
-        uint receivedValue = msg.value;
-        uint startIndex = amountOfPuzzleParts;
+        bytes32 _puzzleDetailsStorageHash = puzzleDetailsStorageHash;
 
-        for (uint i = 0; i < _coins.length; i++) {
-            puzzleParts[startIndex + i] = PuzzlePart({
-                coin: _coins[i],
-                upperBound: _upperBounds[i],
-                commitment: "",
-                solution: "",
-                paidOut: false
-            });
-            receivedValue -= _coins[i];
+        uint256 receivedValue = msg.value;
+
+        for (uint256 i = _coins.length; i > 0; i--) {
+                uint256 index = i - 1;
+
+                emit Initialized(index, _coins[index], _upperBounds[index], _puzzleDetailsStorageHash);
+
+                _puzzleDetailsStorageHash = _hashPuzzleDetails(_coins[index], _upperBounds[index], _puzzleDetailsStorageHash);
+                receivedValue -= _coins[index];
+
         }
 
         require(receivedValue == 0, "The total value sent should be equal to the sum of the coins.");
 
         amountOfPuzzleParts += _coins.length;
+        puzzleDetailsStorageHash = _puzzleDetailsStorageHash;
     }
 
-
-    function commitments() public view returns (bytes[] memory) {
-        bytes[] memory commitments = new bytes[](amountOfPuzzleParts);
-        for (uint i = 0; i < amountOfPuzzleParts; i++) {
-            commitments[i] = puzzleParts[i].commitment;
-        }
-        return commitments;
-    }
-
-    function setCommitments(bytes[] calldata _commitments, uint startIndex) public onlyHelper {
+    function setCommitments(bytes32[] calldata _commitments) public onlyHelper {
         // Change status to SettingCommitments on the first call
         if (contractStatus == Status.Setup) {
             contractStatus = Status.SettingCommitments;
         }
 
         require(contractStatus == Status.SettingCommitments, "Contract is not in SettingCommitments status.");
-        require(startIndex + _commitments.length <= amountOfPuzzleParts, "Commitments exceed the number of puzzle parts.");
+        require(_commitments.length == amountOfPuzzleParts, "The length of the commitments array should be equal to the amount of puzzle parts.");
 
-        // Check that the value at startIndex-1 is not 0 unless startIndex is 0
-        if (startIndex > 0) {
-            require(bytes(puzzleParts[startIndex - 1].commitment).length != 0, "Previous puzzle part commitment must be set.");
+        bytes32 _puzzleCommitmentStorageHash = puzzleCommitmentStorageHash;
+        for (uint256 i = _commitments.length; i > 0; i--) {
+            uint256 index = i - 1;
+
+            emit CommitmentSet(index, _commitments[index], _puzzleCommitmentStorageHash);
+            _puzzleCommitmentStorageHash = _hashPuzzleCommitment(_commitments[index], _puzzleCommitmentStorageHash);
         }
 
-        for (uint i = 0; i < _commitments.length; i++) {
-            uint index = startIndex + i;
-            require(index < amountOfPuzzleParts, "Index out of bounds.");
-            require(bytes(puzzleParts[index].commitment).length == 0, "Commitment has already been set.");
-            puzzleParts[index].commitment = _commitments[i];
-        }
-
-        // Check if all commitments have been set
-        if (startIndex + _commitments.length == amountOfPuzzleParts) {
-            contractStatus = Status.Solving;
-        }
+        puzzleCommitmentStorageHash = _puzzleCommitmentStorageHash;
+        contractStatus = Status.Solving;
     }
 
-    function addSolution(bytes calldata solution, bytes calldata witness) public {
+    function addSolution(bytes calldata solution, bytes calldata witness, bytes32 commitment, bytes32 prevPuzzleCommitmentStorageHash, uint256 coin, uint256 upperBound, bytes32 prevPuzzleDetailsStorageHash) public {
         require(contractStatus == Status.Solving, "Contract is not in Solving status.");
-        require(nextUnsolvedPuzzlePart < amountOfPuzzleParts, "All puzzle parts have already been solved.");
-        require(checkSolution(solution, witness, puzzleParts[nextUnsolvedPuzzlePart].commitment), "The solution is not correct.");
+        require(puzzleDetailsStorageHash != bytes32(0), "All puzzle parts have already been solved.");
+        require(puzzleCommitmentStorageHash != bytes32(0), "All puzzle parts have already been solved.");
 
-        uint256 upperBound = puzzleParts[nextUnsolvedPuzzlePart].upperBound;
+
+        require(puzzleCommitmentStorageHash == _hashPuzzleCommitment(commitment, prevPuzzleCommitmentStorageHash), "The provided puzzle commitment is not correct.");
+        require(checkSolution(solution, witness, commitment), "The solution is not correct.");
+
+        require(puzzleDetailsStorageHash == _hashPuzzleDetails(coin, upperBound, prevPuzzleDetailsStorageHash), "The provided puzzle details are not correct.");
         require(block.timestamp <= startTime + upperBound, "Too late: the time upper bound has been exceeded.");
 
+        emit SolutionReceived(amountOfPuzzleParts, solution, witness);
+        amountOfPuzzleParts--;
+
+        puzzleDetailsStorageHash = prevPuzzleDetailsStorageHash;
+        puzzleCommitmentStorageHash = prevPuzzleCommitmentStorageHash;
+
         // Once the solution is correct, the solver should be paid
-        pay(nextUnsolvedPuzzlePart, msg.sender);
-
-        puzzleParts[nextUnsolvedPuzzlePart].solution = solution;
-
-        nextUnsolvedPuzzlePart++;
+        pay(coin, msg.sender);
     }
 
-    function coins() public view returns (uint[] memory) {
-        uint[] memory coins = new uint[](amountOfPuzzleParts);
-        for (uint i = 0; i < amountOfPuzzleParts; i++) {
-            coins[i] = puzzleParts[i].coin;
-        }
-        return coins;
+    /// We do not need to check if the puzzle part has already been paid out, because after the payment the puzzle details are updated to the next puzzle part.
+    /// This assumes hash functions are collision-resistant.
+    function pay(uint coin, address solver) internal {
+        payable(solver).transfer(coin);
     }
 
-    function upperBounds() public view returns (uint[] memory) {
-        uint[] memory upperBounds = new uint[](amountOfPuzzleParts);
-        for (uint i = 0; i < amountOfPuzzleParts; i++) {
-            upperBounds[i] = puzzleParts[i].upperBound;
-        }
-        return upperBounds;
+    /// If the time upper bound has been exceeded, the owner can call this function to get the money back.
+    /// This way they cancel the latest puzzle part and get their money back.
+    function payBack(bytes32 commitment, bytes32 prevPuzzleCommitmentStorageHash, uint256 coin, uint256 upperBound, bytes32 prevPuzzleDetailsStorageHash) public onlyOwner {
+        require(puzzleCommitmentStorageHash == _hashPuzzleCommitment(commitment, prevPuzzleCommitmentStorageHash), "The provided puzzle commitment is not correct.");
+        require(puzzleDetailsStorageHash == _hashPuzzleDetails(coin, upperBound, prevPuzzleDetailsStorageHash), "The provided puzzle details are not correct.");
+        require(block.timestamp > startTime + upperBound, "Too late: the time upper bound has been exceeded.");
+
+        puzzleDetailsStorageHash = prevPuzzleDetailsStorageHash;
+        puzzleCommitmentStorageHash = prevPuzzleCommitmentStorageHash;
+        payable(owner).transfer(coin);
     }
 
-    function getCommitmentAt(uint puzzlePartIndex) public view returns (bytes memory) {
-        require(puzzlePartIndex < amountOfPuzzleParts, "The puzzle part index is out of bounds.");
-        require(puzzleParts[puzzlePartIndex].commitment.length > 0, "The commitment is not set yet.");
-        return puzzleParts[puzzlePartIndex].commitment;
-    }
-
-    function getSolutionAt(uint puzzlePartIndex) public returns (bytes memory) {
-        require(puzzlePartIndex < amountOfPuzzleParts, "The puzzle part index is out of bounds.");
-
-        bytes memory solutionMemory = puzzleParts[puzzlePartIndex].solution;
-
-        return solutionMemory;
-    }
-
-    function getUpperBoundAt(uint puzzlePartIndex) public view returns (uint) {
-        require(puzzlePartIndex < amountOfPuzzleParts, "The puzzle part index is out of bounds.");
-        return puzzleParts[puzzlePartIndex].upperBound;
-    }
-
-    function solutions() public view returns (bytes[] memory) {
-        bytes[] memory solutions = new bytes[](amountOfPuzzleParts);
-        for (uint i = 0; i < amountOfPuzzleParts; i++) {
-            solutions[i] = puzzleParts[i].solution;
-        }
-        return solutions;
-    }
-
-    function verifySolution(uint puzzlePartIndex) public view returns (bool) {
-        require(puzzlePartIndex < amountOfPuzzleParts, "The puzzle part index is out of bounds.");
-        return puzzleParts[puzzlePartIndex].solution.length > 0;
-    }
-
-    function pay(uint puzzlePartIndex, address solver) internal {
-        require(!puzzleParts[puzzlePartIndex].paidOut, "The puzzle part has already been paid out.");
-        puzzleParts[puzzlePartIndex].paidOut = true;
-        payable(solver).transfer(puzzleParts[puzzlePartIndex].coin);
-    }
-
-    function payBack(uint puzzlePartIndex) public onlyOwner {
-        require(!puzzleParts[puzzlePartIndex].paidOut, "The puzzle part has already been paid out.");
-        puzzleParts[puzzlePartIndex].paidOut = true;
-        payable(owner).transfer(address(this).balance);
-    }
-
-    function checkSolution(bytes calldata solution, bytes calldata witness, bytes memory commitment) public pure returns (bool) {
+    function checkSolution(bytes calldata solution, bytes calldata witness, bytes32 commitment) public pure returns (bool) {
         bytes memory concatenated = abi.encodePacked(solution, witness);
         bytes32 hash = keccak256(concatenated);
 
-        return (hash == abi.decode(commitment, (bytes32)));
+        return (hash == commitment);
     }
+
+    function _hashPuzzleDetails(uint256 coin, uint256 upperBound, bytes32 prevHash) public pure returns (bytes32) {
+        return keccak256(abi.encodePacked(coin, upperBound, prevHash));
+    }
+
+    function _hashPuzzleCommitment(bytes32 commitment, bytes32 prevHash) public pure returns (bytes32) {
+        return keccak256(abi.encodePacked(commitment, prevHash));
+    }
+
 }