Optimised storage by using the hash chain approach, resulting in about 10x performance improvement for setting up the puzzle.

```
1 instances of 1 seconds each
Gas used for initializing: 124768
Gas used for setting commitments: 80407
Gas used for adding solution: 69256
10 instances of 1 seconds each
Gas used for initializing: 171444
Gas used for setting commitments: 114253
Gas used for adding solution: 70010
100 instances of 1 seconds each
Gas used for initializing: 579667
Gas used for setting commitments: 423097
Gas used for adding solution: 70022
1000 instances of 0.01 seconds each
Gas used for initializing batch: 4821036
Gas used for setting commitments: 3625992
Gas used for adding solution: 70034
10000 instances of 0.01 seconds each
Gas used for initializing batch: 48327572
Gas used for setting commitments: 35693543
Gas used for adding solution: 70034
```